requirements.txt
Flask
azure-ai-inference
azure-core